fleshed out form stuff
